+++ Movie List/MovieListTableTableViewController.swift
@@ -1,0 +1,57 @@
+//
+//  MovieListTableTableViewController.swift
+//  Movie List
+//
+//  Created by Marcus Campbell on 5/30/20.
+//  Copyright © 2020 Lambda School. All rights reserved.
+//
+
+import UIKit
+
+class MovieListTableTableViewController: UITableViewController {
+
+    @IBOutlet var ListofMovies: UITableView!
+    
+    
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Uncomment the following line to preserve selection between presentations
+        // self.clearsSelectionOnViewWillAppear = false
+
+        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
+        // self.navigationItem.rightBarButtonItem = self.editButtonItem
+    }
+
+    // MARK: - Table view data source
+
+    override func numberOfSections(in tableView: UITableView) -> Int {
+        // #warning Incomplete implementation, return the number of sections
+        return 0
+    }
+
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        // #warning Incomplete implementation, return the number of rows
+        return 0
+    }
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+}
+
+extension MovieListTableTableViewController: UITableViewDelegate {
+    
+}
+
+extension MovieListTableTableViewController: UITableViewDataSource {
+    
+}
+
