+++ Movie List/MovieListViewController.swift
@@ -1,0 +1,57 @@
+//
+//  MovieListTableTableViewController.swift
+//  Movie List
+//
+//  Created by Marcus Campbell on 5/30/20.
+//  Copyright © 2020 Lambda School. All rights reserved.
+//
+
+import UIKit
+
+class MovieListTableViewController: UIViewController {
+
+    @IBOutlet var ListofMovies: UITableView!
+    
+    var movieList = [Movie(name: "Batman", seen: false)]
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        ListofMovies.delegate = self
+        ListofMovies.dataSource = self
+
+        // Uncomment the following line to preserve selection between presentations
+        // self.clearsSelectionOnViewWillAppear = false
+
+        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
+        // self.navigationItem.rightBarButtonItem = self.editButtonItem
+    }
+
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+}
+
+extension MovieListTableViewController: UITableViewDelegate {
+    
+}
+
+extension MovieListTableViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return movieList.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        return UITableViewCell()
+    }
+    
+    
+}
+
